Many changes
Added a library called re that checks whether the passwords and emails entered are valid. Also implemented a condition on the signup_page that checks whether an email already exists in our database and hence validates that or not. I also implemented a widget called snack bar that informs the user of successful submission